requirements.txt
asgiref==3.8.1
Django==5.2.9
django-cors-headers==4.3.1
django-filter==24.2
django-sslserver==0.22
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
gunicorn==22.0.0
Pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
whitenoise==6.6.0
python-dateutil==2.9.0
tzdata==2024.1
